NGrinder is only compatible with JAR plugins
